perf(web): skip unchanged renders and split oversized editor UI
Add debounce+max-wait preview scheduling and fingerprint skip in renderStore.
Extract scroll-sync/post-tree helpers with tests; cover API origin/sanitize.
Extend type-check to vscode/mcp; split Footer and CustomComponent form panels.
