+++ apps/api/package.json
@@ -9,7 +9,9 @@
     "deploy": "wrangler deploy --minify",
     "db:migrate:local": "wrangler d1 migrations apply md-sync --local",
     "db:migrate:remote": "wrangler d1 migrations apply md-sync --remote",
-    "type-check": "tsc --noEmit"
+    "type-check": "tsc --noEmit",
+    "test": "vitest run",
+    "test:watch": "vitest"
   },
   "dependencies": {
     "hono": "^4.12.28"
@@ -18,6 +20,7 @@
     "@cloudflare/workers-types": "5.20260710.1",
     "@types/node": "^26.1.1",
     "typescript": "~6.0.3",
+    "vitest": "^4.1.10",
     "wrangler": "^4.110.0"
   }
 }
