+++ apps/api/vitest.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  test: {
+    environment: `node`,
+    include: [`src/**/*.test.ts`],
+  },
+})
