+++ apps/web/src/components/editor/Footer.vue
@@ -2,7 +2,9 @@
 import type { MarkdownHeading } from '@/lib/markdown/headings'
 import { StateEffect } from '@codemirror/state'
 import { EditorView } from '@codemirror/view'
-import { ArrowUpDown, BookOpen, ChevronRight, ChevronsUpDown, Clock, Columns2, Ellipsis, Eye, FileText, Keyboard, ListTree, LogIn, Monitor, Moon, PenLine, Pilcrow, Search, Share2, Smartphone, Sun, Type, User } from '@lucide/vue'
+import { ArrowUpDown, BookOpen, Clock, Columns2, Ellipsis, Eye, FileText, Keyboard, LogIn, Monitor, Moon, PenLine, Pilcrow, Share2, Smartphone, Sun, Type, User } from '@lucide/vue'
+import FooterDocumentSwitcher from '@/components/editor/footer/FooterDocumentSwitcher.vue'
+import FooterOutlinePopover from '@/components/editor/footer/FooterOutlinePopover.vue'
 import {
   Popover,
   PopoverContent,
@@ -19,9 +21,7 @@ import { getLocaleOption, getNextLocale } from '@/i18n/constants'
 import { formatRelativeTime } from '@/lib/format/relative-time'
 import {
   clampGoToLineValue,
-  findOutlineFocusIndex,
   jumpToLine,
-  moveOutlineFocusIndex,
 } from '@/lib/markdown/headingNavigation'
 import { computeHeadingBreadcrumbs, extractMarkdownHeadings } from '@/lib/markdown/headings'
 import { isAccountUiEnabled } from '@/services/account/config'
@@ -95,76 +95,6 @@ function toggleLanguage() {
   localeStore.cycleLocale()
 }
 
-// 快速文档切换器
-const isSwitcherOpen = ref(false)
-const switcherQuery = ref(``)
-const switcherInputRef = ref<HTMLInputElement | null>(null)
-
-interface TreeNode {
-  id: string
-  title: string
-  updateDatetime: Date
-  depth: number
-  children: TreeNode[]
-}
-
-function buildTree(posts: typeof postStore.posts): TreeNode[] {
-  const map = new Map<string, TreeNode>()
-  for (const p of posts) {
-    map.set(p.id, { id: p.id, title: p.title, updateDatetime: p.updateDatetime, depth: 0, children: [] })
-  }
-  const roots: TreeNode[] = []
-  for (const p of posts) {
-    const node = map.get(p.id)
-    if (!node)
-      continue
-    if (p.parentId && map.has(p.parentId)) {
-      map.get(p.parentId)!.children.push(node)
-    }
-    else {
-      roots.push(node)
-    }
-  }
-  // 排序子节点
-  function sortChildren(nodes: TreeNode[]) {
-    nodes.sort((a, b) => new Date(b.updateDatetime).getTime() - new Date(a.updateDatetime).getTime())
-    nodes.forEach(n => sortChildren(n.children))
-  }
-  sortChildren(roots)
-  return roots
-}
-
-function flattenTree(nodes: TreeNode[], depth = 0): Array<{ id: string, title: string, updateDatetime: Date, depth: number }> {
-  const result: Array<{ id: string, title: string, updateDatetime: Date, depth: number }> = []
-  for (const node of nodes) {
-    result.push({ id: node.id, title: node.title, updateDatetime: node.updateDatetime, depth })
-    if (node.children.length)
-      result.push(...flattenTree(node.children, depth + 1))
-  }
-  return result
-}
-
-const flatPosts = computed(() => flattenTree(buildTree(postStore.posts)))
-
-const filteredPosts = computed(() => {
-  const q = switcherQuery.value.toLowerCase().trim()
-  if (!q)
-    return flatPosts.value
-  return flatPosts.value.filter(p => p.title.toLowerCase().includes(q))
-})
-
-function openSwitcher() {
-  isSwitcherOpen.value = true
-  switcherQuery.value = ``
-  nextTick(() => switcherInputRef.value?.focus())
-}
-
-function switchToPost(id: string) {
-  postStore.currentPostId = id
-  isSwitcherOpen.value = false
-  nextTick(() => editor.value?.focus())
-}
-
 // 光标位置
 const cursorLine = ref(1)
 const cursorCol = ref(1)
@@ -203,9 +133,6 @@ const attachedViews = new WeakSet()
 // 大纲 & 面包屑
 const breadcrumbs = ref<MarkdownHeading[]>([])
 const allHeadings = ref<MarkdownHeading[]>([])
-const isOutlineOpen = ref(false)
-const outlineScrollRef = ref<HTMLElement | null>(null)
-const outlineFocusIndex = ref(-1)
 
 function updateCursorInfo(view: EditorView, options: { rebuildHeadings?: boolean } = {}) {
   const state = view.state
@@ -260,37 +187,18 @@ const activeHeadingLine = computed(() => {
   return breadcrumbs.value[breadcrumbs.value.length - 1].line
 })
 
-function getOutlineLevelClass(level: number) {
-  if (level === 1)
-    return `font-semibold text-foreground`
-  if (level === 2)
-    return `font-medium text-foreground/90`
-  return `text-muted-foreground`
-}
-
-function jumpToHeadingAndClose(line: number) {
-  jumpToHeading(line)
-  isOutlineOpen.value = false
-}
-
-watch(isOutlineOpen, (open) => {
-  if (open) {
-    syncOutlineFocusIndex()
-    nextTick(() => {
-      outlineScrollRef.value?.focus()
-      const el = outlineScrollRef.value?.querySelector(`[data-outline-index="${outlineFocusIndex.value}"]`)
-        ?? outlineScrollRef.value?.querySelector(`[data-active="true"]`)
-      el?.scrollIntoView({ block: `nearest` })
-    })
-  }
-})
-
 function jumpToHeading(line: number) {
   const view = editor.value
   if (!view)
     return
   jumpToLine(view as EditorView, line)
   updateCursorInfo(view as EditorView)
+  nextTick(() => editor.value?.focus())
+}
+
+function switchToPost(id: string) {
+  postStore.currentPostId = id
+  nextTick(() => editor.value?.focus())
 }
 
 watch(() => uiStore.goToLineRequest, () => {
@@ -299,51 +207,6 @@ watch(() => uiStore.goToLineRequest, () => {
   openGoToLine()
 })
 
-function syncOutlineFocusIndex() {
-  outlineFocusIndex.value = findOutlineFocusIndex(allHeadings.value, activeHeadingLine.value)
-}
-
-function scrollOutlineFocusIntoView() {
-  nextTick(() => {
-    const container = outlineScrollRef.value
-    if (!container || outlineFocusIndex.value < 0)
-      return
-    const el = container.querySelector(`[data-outline-index="${outlineFocusIndex.value}"]`)
-    el?.scrollIntoView({ block: `nearest` })
-  })
-}
-
-function handleOutlineKeydown(event: KeyboardEvent) {
-  if (allHeadings.value.length === 0)
-    return
-
-  const key = event.key
-  if (key === `Escape`) {
-    isOutlineOpen.value = false
-    editor.value?.focus()
-    event.preventDefault()
-    return
-  }
-
-  if (key === `Enter`) {
-    const item = allHeadings.value[outlineFocusIndex.value]
-    if (item)
-      jumpToHeadingAndClose(item.line)
-    event.preventDefault()
-    return
-  }
-
-  if (key === `ArrowUp` || key === `ArrowDown` || key === `Home` || key === `End`) {
-    outlineFocusIndex.value = moveOutlineFocusIndex(
-      allHeadings.value,
-      outlineFocusIndex.value,
-      key,
-    )
-    scrollOutlineFocusIntoView()
-    event.preventDefault()
-  }
-}
-
 // 上次保存时间
 const savedTimeAgo = computed(() => {
   void locale.value
@@ -351,11 +214,6 @@ const savedTimeAgo = computed(() => {
     return ``
   return formatRelativeTime(currentPost.value.updateDatetime)
 })
-
-function relativeTime(datetime: string | Date) {
-  void locale.value
-  return formatRelativeTime(datetime)
-}
 
 // 每 10 秒刷新一次相对时间（页面不可见时暂停）
 const refreshKey = ref(0)
@@ -449,128 +307,18 @@ const showDeviceToggle = computed(() => viewMode.value !== `edit` && !isMobile.v
       </div>
 
       <!-- 文档切换器 -->
-      <Popover v-model:open="isSwitcherOpen">
-        <PopoverTriggerPrimitive as-child>
-          <Tooltip>
-            <TooltipTrigger as-child>
-              <button
-                class="ml-1.5 flex max-w-24 shrink-0 cursor-pointer items-center gap-1 rounded px-1.5 py-0.5 transition-colors hover:bg-accent hover:text-foreground sm:ml-2 sm:max-w-36"
-                @click="openSwitcher"
-              >
-                <FileText class="size-3 shrink-0 opacity-60" />
-                <span class="truncate">{{ currentPost?.title || t('common.unnamed') }}</span>
-                <ChevronsUpDown class="size-3 shrink-0 opacity-40" />
-              </button>
-            </TooltipTrigger>
-            <TooltipContent v-if="!isSwitcherOpen" side="top" :side-offset="6" class="text-xs text-muted-foreground">
-              <p>{{ t('footer.switchDocument') }}</p>
-            </TooltipContent>
-          </Tooltip>
-        </PopoverTriggerPrimitive>
-        <PopoverContent side="top" :side-offset="8" align="start" class="w-64 p-0">
-          <div class="flex items-center gap-2 border-b px-3 py-2">
-            <Search class="size-3.5 shrink-0 opacity-50" />
-            <input
-              ref="switcherInputRef"
-              v-model="switcherQuery"
-              type="text"
-              class="h-5 w-full bg-transparent text-xs text-foreground outline-none placeholder:text-muted-foreground"
-              :placeholder="t('footer.searchDocuments')"
-              @keydown.escape="isSwitcherOpen = false"
-            >
-          </div>
-          <div class="max-h-52 overflow-y-auto py-1">
-            <div v-if="filteredPosts.length === 0" class="px-3 py-4 text-center text-xs text-muted-foreground">
-              {{ t('footer.noMatchingDocuments') }}
-            </div>
-            <button
-              v-for="post in filteredPosts"
-              :key="post.id"
-              class="flex w-full cursor-pointer items-center gap-2 py-1.5 pr-3 text-left text-xs transition-colors hover:bg-accent"
-              :class="post.id === postStore.currentPostId ? 'bg-accent/50 text-foreground' : 'text-muted-foreground'"
-              :style="{ paddingLeft: `${12 + post.depth * 16}px` }"
-              @click="switchToPost(post.id)"
-            >
-              <FileText class="size-3 shrink-0 opacity-50" />
-              <span class="min-w-0 flex-1 truncate">{{ post.title }}</span>
-              <span class="shrink-0 text-[10px] opacity-50">{{ relativeTime(post.updateDatetime) }}</span>
-            </button>
-          </div>
-        </PopoverContent>
-      </Popover>
+      <FooterDocumentSwitcher
+        :current-title="currentPost?.title"
+        @select="switchToPost"
+      />
 
       <!-- 大纲视图 -->
-      <Popover v-model:open="isOutlineOpen">
-        <PopoverTriggerPrimitive as-child>
-          <Tooltip>
-            <TooltipTrigger as-child>
-              <button
-                class="ml-1 flex min-w-0 cursor-pointer items-center gap-1.5 rounded px-1 py-0.5 transition-colors hover:bg-accent hover:text-foreground sm:mx-3 sm:px-1.5"
-                @click="isOutlineOpen = !isOutlineOpen"
-              >
-                <ListTree class="size-3.5 shrink-0 opacity-60 sm:size-3" />
-                <div v-if="breadcrumbs.length" class="hidden min-w-0 items-center gap-0.5 truncate sm:flex">
-                  <template v-for="(crumb, idx) in breadcrumbs" :key="crumb.line">
-                    <ChevronRight v-if="idx > 0" class="size-3 shrink-0 opacity-30" />
-                    <span class="max-w-24 truncate">{{ crumb.title }}</span>
-                  </template>
-                </div>
-                <span v-else class="hidden opacity-50 sm:inline">{{ t('footer.outline') }}</span>
-              </button>
-            </TooltipTrigger>
-            <TooltipContent v-if="!isOutlineOpen" side="top" :side-offset="6" class="text-xs text-muted-foreground">
-              <p>{{ t('footer.outlineTooltip') }}</p>
-            </TooltipContent>
-          </Tooltip>
-        </PopoverTriggerPrimitive>
-        <PopoverContent
-          side="top"
-          :side-offset="8"
-          align="center"
-          class="w-72 p-0"
-        >
-          <div class="flex items-center justify-between border-b px-3 py-2">
-            <span class="text-xs font-medium tracking-wide text-muted-foreground">{{ t('footer.outline') }}</span>
-            <span class="text-[10px] tabular-nums text-muted-foreground/50">{{ t('footer.headingCount', { count: allHeadings.length }) }}</span>
-          </div>
-          <div
-            ref="outlineScrollRef"
-            tabindex="-1"
-            class="max-h-72 overflow-y-auto overflow-x-hidden py-1 outline-none"
-            @keydown.stop="handleOutlineKeydown"
-          >
-            <template v-if="allHeadings.length > 0">
-              <button
-                v-for="(item, index) in allHeadings"
-                :key="item.line"
-                :data-outline-index="index"
-                :data-active="activeHeadingLine === item.line"
-                class="group flex w-full items-start px-2 py-1 text-left text-[13px] transition-colors hover:bg-accent"
-                :class="[
-                  activeHeadingLine === item.line ? 'bg-accent/60' : '',
-                  outlineFocusIndex === index ? 'bg-accent/40 ring-1 ring-primary/30' : '',
-                ]"
-                :style="{ paddingLeft: `${8 + (item.level - 1) * 16}px` }"
-                @click="jumpToHeadingAndClose(item.line)"
-              >
-                <span
-                  class="mr-2 mt-[7px] inline-block size-1 shrink-0 rounded-full bg-current transition-opacity"
-                  :class="activeHeadingLine === item.line ? 'opacity-80' : item.level <= 2 ? 'opacity-40' : 'opacity-20'"
-                />
-                <span
-                  class="line-clamp-1 leading-relaxed"
-                  :class="getOutlineLevelClass(item.level)"
-                >
-                  {{ item.title }}
-                </span>
-              </button>
-            </template>
-            <div v-else class="px-3 py-8 text-center text-xs text-muted-foreground">
-              {{ t('footer.noHeadings') }}
-            </div>
-          </div>
-        </PopoverContent>
-      </Popover>
+      <FooterOutlinePopover
+        :breadcrumbs="breadcrumbs"
+        :headings="allHeadings"
+        :active-heading-line="activeHeadingLine"
+        @jump="jumpToHeading"
+      />
 
       <!-- 桌面端占位 flex-1 -->
       <div class="hidden min-w-0 flex-1 sm:block" />
