+++ apps/web/src/composables/useScrollSync.ts
@@ -1,11 +1,13 @@
 import type { Text } from '@codemirror/state'
 import type { EditorView } from '@codemirror/view'
 import type { MaybeRefOrGetter, Ref } from 'vue'
-
-interface SourceBlock {
-  startLine: number
-  endLine: number
-}
+import type { SourceBlock } from '@/lib/preview/scroll-sync-blocks'
+import {
+  mapBlockIndex,
+  parseSourceBlocks,
+  sourceBlockIndexForLine,
+  sourceLineForBlockIndex,
+} from '@/lib/preview/scroll-sync-blocks'
 
 /**
  * 同步滚动：编辑器滚动时同步预览区，预览区滚动时同步编辑器。
@@ -35,26 +37,6 @@ export function useScrollSync(
   // ============================================================
   let cachedDoc: Text | null = null
   let cachedSourceBlocks: SourceBlock[] | null = null
-
-  function parseSourceBlocks(doc: Text): SourceBlock[] {
-    const blocks: SourceBlock[] = []
-    let blockStart = -1
-
-    for (let i = 1; i <= doc.lines; i++) {
-      const isEmpty = doc.line(i).text.trim() === ''
-      if (!isEmpty && blockStart === -1) {
-        blockStart = i
-      }
-      else if (isEmpty && blockStart !== -1) {
-        blocks.push({ startLine: blockStart, endLine: i - 1 })
-        blockStart = -1
-      }
-    }
-    if (blockStart !== -1) {
-      blocks.push({ startLine: blockStart, endLine: doc.lines })
-    }
-    return blocks
-  }
 
   function getSourceBlocks(doc: Text): SourceBlock[] {
     if (cachedDoc === doc && cachedSourceBlocks)
@@ -146,44 +128,6 @@ export function useScrollSync(
     cachedPreviewBlocks = blocks
     cachedPreviewOffsetTops = offsetTops
     return { blocks, offsetTops }
-  }
-
-  // ============================================================
-  // 映射辅助
-  // ============================================================
-  function sourceBlockIndexForLine(
-    sourceBlocks: SourceBlock[],
-    lineNo: number,
-  ): number {
-    for (let i = sourceBlocks.length - 1; i >= 0; i--) {
-      if (sourceBlocks[i].startLine <= lineNo)
-        return i
-    }
-    return 0
-  }
-
-  function sourceLineForBlockIndex(
-    sourceBlocks: SourceBlock[],
-    blockIndex: number,
-  ): number {
-    if (sourceBlocks.length === 0)
-      return 1
-    const idx = Math.min(blockIndex, sourceBlocks.length - 1)
-    return sourceBlocks[idx].startLine
-  }
-
-  /**
-   * 按比例将源块索引映射到预览块索引。
-   * 用 Math.round 使映射更精准。
-   */
-  function mapBlockIndex(
-    index: number,
-    sourceTotal: number,
-    targetTotal: number,
-  ): number {
-    if (sourceTotal <= 1 || targetTotal <= 1)
-      return Math.round((index / Math.max(sourceTotal - 1, 1)) * Math.max(targetTotal - 1, 1))
-    return Math.round((index / (sourceTotal - 1)) * (targetTotal - 1))
   }
 
   // ============================================================
