+++ package.json
@@ -28,8 +28,8 @@
     "lint": "eslint . --fix",
     "type-check": "pnpm type-check:web && pnpm type-check:packages",
     "type-check:web": "pnpm --filter @md/web exec vue-tsc --noEmit",
-    "type-check:packages": "pnpm --filter @md/core --filter @md/shared --filter @md/api run type-check",
-    "test": "pnpm --filter @md/core test && pnpm --filter @md/web test",
+    "type-check:packages": "pnpm --filter @md/core --filter @md/shared --filter @md/api --filter @md/mcp-server run type-check && pnpm --prefix ./apps/vscode run type-check",
+    "test": "pnpm --filter @md/core test && pnpm --filter @md/web test && pnpm --filter @md/api test",
     "postinstall": "simple-git-hooks",
     "inspector": "pnpx node-modules-inspector",
     "utools:package": "node ./scripts/package-utools.mjs",
